Jflynn129 dev (#2)
* Initial commit
Make WNC driver as a seperate driver
fixing mbed_lib.json, was missing a closing '}'
Adding the BuffereSerial library to the wnc14a2a-driver because this driver requires it.  The wnc14a2a communicates via the UART which uses this library.
Modified the time-out operation. It is possible the caller called RX to simply monitor for unsolicited incomming messages, so don't fail on TIMEOUT, simply return any data that did come in and keep looking for more
stoped any pending read queued events in the case a user calls to close the socket while one is pending.
rewrite receive handler so that it handles the case when 'bind' has been called.
Switched from BufferedSerial to UARTSerial because UARTSerial is supported in mbed-os while BufferedSerial is not.  Added a new WncIO class (WNCIO.h) to handle uart i/o to minimize impacting existing code.
Changed pin names from K64F style to Adrunio style for cross platform compiles.
* add README file, modify initialization for static instantiation